backend/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-jsonapi==2.8.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.1
inflection==0.3.1
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
